backend/requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
loguru==0.6.0
Mako==1.1.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
loguru==0.6.0
Mako==1.1.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
psycopg==3.0.9
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
loguru==0.6.0
Mako==1.1.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
psycopg==3.0.9
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
